Fix scoring residences

+++ SettlementSimulation.Engine/Interfaces/IBuildingFitness.cs
@@ -5,7 +5,7 @@ namespace SettlementSimulation.Engine.Interfaces
     public interface IBuildingFitness
     {
         void SetFitness(BuildingRule model);
-        int CalculateFitness(BuildingRule model);
-        int Fitness { get; set; }
+        double CalculateFitness(BuildingRule model);
+        double Fitness { get; set; }
     }
 }
